requirements.txt
atomInSmiles==1.0.2
datasets==2.19.2
einops==0.8.1
fsspec==2024.3.0
google_auth==2.38.0
jsonargparse[signatures]==4.35.0
lightning==2.5.0.post0
lightning-utilities==0.11.9
matplotlib==3.10.5
numpy==1.26.4
openbabel-wheel==3.1.1.22
packaging==24.2
pandas==2.3.1
pip==24.2
Pillow==11.3.0
psutil==6.1.1
pytdc==1.1.15
PyYAML==6.0.2
rdkit==2023.9.6
Requests==2.32.4
scikit_learn==1.2.2
scipy==1.15.0
seaborn==0.13.2
setuptools==75.6.0
tokenizers==0.21.4
torch==2.5.1
torchdata==0.10.1
torchmetrics==1.6.1
tqdm==4.67.1
transformers==4.47.1
wandb==0.21.1
